container: Support recreate from private registry
