+++ bin/container/create/index.js
@@ -1,11 +1,7 @@
 'use strict';
 
 const Cli = require('lib/cli');
-const os = require('os');
-const path = require('path');
-const fs = require('fs');
-const util = require('util');
-const readFile = util.promisify(fs.readFile);
+const registry = require('../registry');
 
 const options = {
     name: {
@@ -22,22 +18,6 @@ const options = {
         description: 'Container type',
         type: 'string',
         required: true,
-    },
-    ['registry-username']: {
-        description: 'Username to access container registry',
-        type: 'string',
-        required: false,
-    },
-    ['registry-password']: {
-        description: 'Username to access container registry',
-        type: 'string',
-        required: false,
-    },
-    ['registry-dockercfg']: {
-        description: 'Use credentials from .dockercfg',
-        type: 'boolean',
-        required: false,
-        defaultValue: false,
     },
     expose: {
         description: 'Mapping port to expose to the world as external:internal',
@@ -67,69 +47,22 @@ const options = {
     },
 };
 
-const config_files = [
-    path.join(process.cwd(), '.dockercfg'),
-    path.join(os.homedir(), '.dockercfg'),
-    path.join(os.homedir(), '.docker/config.json'),
-];
-
-const getRegistryName = args => {
-    const parts = args.image.split('/');
-    if (parts.length > 1) {
-        return parts[0];
-    }
-};
-const getRegistry = async args => {
-    const registry_name = getRegistryName(args);
-
-    if (!registry_name) {
-        return;
-    }
-
-    if (args['registry-username']) {
-        return {
-            username: args['registry-username'],
-            password: args['registry-password'],
-        };
-    }
-
-    if (!args['registry-dockercfg']) {
-        return;
-    }
-
-    for (const config of config_files) {
-        let content;
-        try {
-            content = await readFile(config, {encoding: 'utf-8'});
-        } catch (err) {
-            continue;
-        }
-        const dockercfg = JSON.parse(content);
-
-        if (!dockercfg.auths || !dockercfg.auths[registry_name] || !dockercfg.auths[registry_name].auth) continue;
-
-        return {
-            token: dockercfg.auths[registry_name].auth,
-        };
-    }
-
-};
 
 module.exports = resource => Cli.createCommand('create', {
     description: `Create ${resource.title}`,
     plugins: resource.plugins,
     genericOptions: ['tag'],
     priority: 25,
-    options: Object.assign({}, options, resource.options),
+    options: Object.assign({}, options, resource.options, registry.options),
     dirname: __dirname,
     handler: async args => {
-        const registry = await getRegistry(args);
 
         const body = {
             name: args.name,
             image: args.image,
             service: args.type,
             env: args.env,
+            registry: await registry.parseArgs(args),
             volumes: args.volume.map(v => {
                 const [sourceFull, target] = v.split(':');
                 const [source, ...sourcePath] = sourceFull.split('/');
@@ -151,10 +84,6 @@ module.exports = resource => Cli.createCommand('create', {
             }),
             tag: require('lib/tags').createTagObject(args.tag),
         };
-
-        if (registry) {
-            body.registry = registry;
-        }
 
         if (args.command) {
             body.command = args.command;
